V course/Neural networks/IV lab/requirements.txt
torchvision
numpy
torch
pillow
tqdm
matplotlib